+++ package.json
@@ -25,14 +25,13 @@
     "@nestjs/config": "^2.3.0",
     "@nestjs/core": "^9.0.0",
     "@nestjs/platform-express": "^9.0.0",
+    "@nestjs/schedule": "^2.2.0",
     "@nestjs/swagger": "^6.2.1",
     "@nestjs/typeorm": "^9.0.1",
     "bcrypt": "^5.1.0",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
-    "date-fns": "^2.29.3",
     "hashids": "^2.2.10",
-    "moment": "^2.29.4",
     "pg": "^8.9.0",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
@@ -43,6 +42,7 @@
     "@nestjs/cli": "^9.0.0",
     "@nestjs/schematics": "^9.0.0",
     "@nestjs/testing": "^9.0.0",
+    "@types/cron": "^2.0.0",
     "@types/express": "^4.17.13",
     "@types/hashids": "^2.0.1",
     "@types/jest": "28.1.8",
