+++ src/tokens/dto/save-url.dto.ts
@@ -1,5 +1,4 @@
-import { ApiProperty } from '@nestjs/swagger';
-import { ApiPropertyOptional } from '@nestjs/swagger/dist';
+import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
 import { IsOptional, IsUrl } from 'class-validator';
 import { IsDate, IsInt, IsPositive, IsStrongPassword } from 'class-validator';
