+++ src/tokens/tokens.module.ts
@@ -1,0 +1,13 @@
+import { Module } from '@nestjs/common';
+import { TokensController } from './tokens.controller';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { Tokens } from './entities/tokens.entity';
+import { TokensService } from './tokens.service';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([Tokens])],
+  controllers: [TokensController],
+  providers: [TokensService],
+  exports: [TokensService],
+})
+export class TokensModule {}
